Video analysis - change average count from float to int

+++ imageai/Detection/Custom/__init__.py
@@ -1121,8 +1121,7 @@ class CustomVideoObjectDetection:
                                         this_second_counting[eachItem] = eachCountingDict[eachItem]
 
                             for eachCountingItem in this_second_counting:
-                                this_second_counting[eachCountingItem] = this_second_counting[
-                                                                             eachCountingItem] / frames_per_second
+                                this_second_counting[eachCountingItem] = int(this_second_counting[eachCountingItem] / frames_per_second)
 
                             if (return_detected_frame == True):
                                 per_second_function(int(counting / frames_per_second),
@@ -1156,9 +1155,7 @@ class CustomVideoObjectDetection:
                                         this_minute_counting[eachItem] = eachCountingDict[eachItem]
 
                             for eachCountingItem in this_minute_counting:
-                                this_minute_counting[eachCountingItem] = this_minute_counting[
-                                                                             eachCountingItem] / (
-                                                                                 frames_per_second * 60)
+                                this_minute_counting[eachCountingItem] = int(this_minute_counting[eachCountingItem] / (frames_per_second * 60))
 
                             if (return_detected_frame == True):
                                 per_minute_function(int(counting / (frames_per_second * 60)),
